+++ src/edge.rs
@@ -40,24 +40,24 @@ impl Eq for Edge {}
 
 #[derive(Debug, Clone)]
 pub struct WeightedEdge {
-    pub u: f64,
-    pub v: f64,
-    pub weight: f64,
+    pub u: i64,
+    pub v: i64,
+    pub weight: f32,
     pub directed: bool,
 }
 
 impl Default for WeightedEdge {
     fn default() -> Self {
         Self {
-            u: -1.0,
-            v: -1.0,
+            u: -1,
+            v: -1,
             weight: -1.0,
             directed: false,
         }
     }
 }
 impl WeightedEdge {
-    pub fn new(u: f64, v: f64, directed: bool) -> Self {
+    pub fn new(u: i64, v: i64, directed: bool) -> Self {
         let (u, v) = if !directed && u > v { (v, u) } else { (u, v) };
         Self {
             u,
@@ -67,7 +67,7 @@ impl WeightedEdge {
         }
     }
 
-    pub fn new_weighted(u: f64, v: f64, weight: f64, directed: bool) -> Self {
+    pub fn new_weighted(u: i64, v: i64, weight: f32, directed: bool) -> Self {
         let (u, v) = if !directed && u > v { (v, u) } else { (u, v) };
         Self {
             u,
@@ -78,7 +78,7 @@ impl WeightedEdge {
     }
 
     pub fn is_empty(&self) -> bool {
-        self.u == -1.0 && self.v == -1.0
+        self.u == -1 && self.v == -1
     }
 }
