+++ src/node_distance.rs
@@ -1,18 +1,18 @@
 use crate::kdtree::KDTree;
 
-pub fn node_distance(left: &KDTree, right: &KDTree) -> f64 {
+pub fn node_distance(left: &KDTree, right: &KDTree) -> f32 {
     for d in 0..left.dim() {
         if left.get_min(d) > right.get_max(d) || right.get_min(d) > left.get_max(d) {
             let mut rsqr = 0.0;
             for dd in d..left.dim() {
-                let mut tmp = f64::max(
+                let mut tmp = f32::max(
                     left.get_min(dd) - right.get_max(dd),
                     right.get_min(dd) - left.get_max(dd),
                 );
-                tmp = f64::max(tmp, 0.0);
+                tmp = f32::max(tmp, 0.0);
                 rsqr += tmp * tmp;
             }
-            return f64::sqrt(rsqr);
+            return f32::sqrt(rsqr);
         }
     }
     0.0
